% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cran_repo}
\alias{check_cran_repo}
\title{Check if CRAN binary repository is available}
\usage{
check_cran_repo(
  repo_url,
  fallback_version = "4.4",
  target = c("windows", "mac")
)
}
\arguments{
\item{repo_url}{Repository URL}

\item{fallback_version}{Fallback R version if primary fails}

\item{target}{Target platform}
}
\value{
List with repo_url and available packages
}
\description{
Check if CRAN binary repository is available
}
\keyword{internal}
