% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_windows.R
\name{check_r_extraction}
\alias{check_r_extraction}
\title{Check if R extraction was successful}
\usage{
check_r_extraction(path)
}
\arguments{
\item{path}{Path to check}
}
\value{
TRUE if R appears to be extracted
}
\description{
Check if R extraction was successful
}
\keyword{internal}
