% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_system_deps}
\alias{check_system_deps}
\title{Check system dependencies for building}
\usage{
check_system_deps(target = c("windows", "mac"))
}
\arguments{
\item{target}{Target platform ("windows" or "mac")}
}
\value{
Named list of available tools
}
\description{
Check system dependencies for building
}
\keyword{internal}
