% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_app_files}
\alias{copy_app_files}
\title{Copy app files to distribution directory}
\usage{
copy_app_files(project_path, dist_dir, entry_script, extra_dirs = NULL)
}
\arguments{
\item{project_path}{Path to project root}

\item{dist_dir}{Path to distribution directory}

\item{entry_script}{Relative path to entry script from project root}

\item{extra_dirs}{Additional directories to copy}
}
\value{
Invisible TRUE
}
\description{
Copy app files to distribution directory
}
\keyword{internal}
