% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{copy_app_files_electron}
\alias{copy_app_files_electron}
\title{Copy app files for Electron wrapper}
\usage{
copy_app_files_electron(
  project_path,
  app_bundle_path,
  entry_script,
  extra_dirs
)
}
\arguments{
\item{project_path}{Source project path}

\item{app_bundle_path}{Path to the .app bundle}

\item{entry_script}{Entry script name}

\item{extra_dirs}{Extra directories to copy}
}
\description{
Copies Shiny app files to the Electron app bundle's Resources directory.
}
\keyword{internal}
