% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{copy_app_files_mac}
\alias{copy_app_files_mac}
\title{Copy app files to macOS app bundle}
\usage{
copy_app_files_mac(
  project_path,
  dist_dir,
  entry_script,
  extra_dirs = NULL,
  app_name
)
}
\arguments{
\item{project_path}{Path to project root}

\item{dist_dir}{Path to distribution directory}

\item{entry_script}{Relative path to entry script from project root}

\item{extra_dirs}{Additional directories to copy}

\item{app_name}{Name of the app}
}
\value{
Invisible TRUE
}
\description{
Copy app files to macOS app bundle
}
\keyword{internal}
