% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{create_mac_app_launcher}
\alias{create_mac_app_launcher}
\title{Create macOS app launcher and Info.plist}
\usage{
create_mac_app_launcher(dist_dir, entry_script, shiny_command, app_name)
}
\arguments{
\item{dist_dir}{Distribution directory}

\item{entry_script}{Entry script name}

\item{shiny_command}{R command to run}

\item{app_name}{Name of the app}
}
\value{
Invisible path to executable
}
\description{
Create macOS app launcher and Info.plist
}
\keyword{internal}
