% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{create_mac_launcher}
\alias{create_mac_launcher}
\title{Create macOS launcher shell script}
\usage{
create_mac_launcher(dist_dir, entry_script, shiny_command)
}
\arguments{
\item{dist_dir}{Distribution directory}

\item{entry_script}{Entry script name}

\item{shiny_command}{R command to run}
}
\value{
Invisible path to script
}
\description{
Create macOS launcher shell script
}
\keyword{internal}
