% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_windows.R
\name{create_windows_launcher}
\alias{create_windows_launcher}
\title{Create Windows launcher batch file}
\usage{
create_windows_launcher(dist_dir, entry_script, shiny_command)
}
\arguments{
\item{dist_dir}{Distribution directory}

\item{entry_script}{Entry script name}

\item{shiny_command}{R command to run}
}
\value{
Invisible path to batch file
}
\description{
Create Windows launcher batch file
}
\keyword{internal}
