% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_zip_archive}
\alias{create_zip_archive}
\title{Create a zip archive of the distribution}
\usage{
create_zip_archive(dist_dir, output_zip)
}
\arguments{
\item{dist_dir}{Path to distribution directory}

\item{output_zip}{Path to output zip file}
}
\value{
Invisible path to zip file
}
\description{
Create a zip archive of the distribution
}
\keyword{internal}
