% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{crs_corr}
\alias{crs_corr}
\title{Cross-section Correlation}
\usage{
crs_corr(DT, Date, NoUse, Use)
}
\arguments{
\item{DT}{A data.frame or data.table, contains all data.}

\item{Date}{A single character, variable name which indicate cross-section.}

\item{NoUse}{Character vector, some variables we don't care (including all character variables).}

\item{Use}{Character vector, reverse of NoUse, some variables we care. Cannot be used at the
same time as NoUse.}
}
\value{
A matrix-like asymmetric data.table, upper is Spearman Correlation and lower
is Pearson Correlation.
}
\description{
Calculate Cross-section Correlation between variables in a data.table.
}
\details{
Every \strong{Date} (a cross-section) we calculate the Spearman / Pearson correlation
between variables, so we have a time series of correlation. Then we can calculate
the time series' mean.

Warning: because we use \code{stats::cor} to calculate correlations each cross-section,
and we use \code{cor(x, use = "na.or.complete")}, so if there exist some NAs, the result
of NoUse and Use may be quite different.
}
\examples{
data(stock)
crs_corr(DT = stock,
         Date = "month",
         NoUse = c("stkcd", "industry", "year", "factor1", "factor2"))

crs_corr(DT = stock, Date = "month", Use = c("return", "size"))

}
\seealso{
\code{\link[stats]{cor}}

Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
