% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_persist.R
\name{crs_pers}
\alias{crs_pers}
\title{Cross-section Persistence}
\usage{
crs_pers(DT, Date, Id, Var, Lag)
}
\arguments{
\item{DT}{A data.frame or data.table, contains all data.}

\item{Date}{A single character, variable name which indicate cross-section.}

\item{Id}{A single character, variable name which indicate different stocks. Because we should
get the lag variable of different stock, this Id can make sure every stock matchs itself.
If you don't have an Id indicator, just let \code{Id = NULL}}

\item{Var}{Character vector, some variables we want to calculate Cross-section Persistence.}

\item{Lag}{Numeric vector, the lag periods use to compute correlation.}
}
\value{
A data.table contains variables and their persistence.
}
\description{
Calculate some variables' Cross-section Persistence in a data.table.
}
\details{
Every \strong{Date} (a cross-section) we calculate the Pearson correlation between
variable A and its lag A_L_n (n is the lag periods), use different stocks' data.
Finanly we have a time series of correlation. Then we calculate its mean.
}
\examples{
crs_pers(DT = stock,
         Id = "stkcd",
         Date = "month",
         Var = c("return", "size"),
         Lag = c(1, 3, 5))

}
\seealso{
Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
