% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_single.R
\name{crs_port}
\alias{crs_port}
\title{Single Sort Portfolio Analysis}
\usage{
crs_port(
  DT,
  LHS,
  RHS,
  Date,
  Weight = NULL,
  Risk = NULL,
  Newey = T,
  GroupN = 5,
  Detail = F
)
}
\arguments{
\item{DT}{A data.frame or data.table, contains all data.}

\item{LHS}{Single character, the name of outcome / independent variable.}

\item{RHS}{Single character, the name of sort variable.}

\item{Date}{Single character, variable name which indicate cross-section.}

\item{Weight}{Single character, variable name which used to calculate value-weighted mean.}

\item{Risk}{Character vector, risk adjust using factors like Fama-French (MKT, SMB, HML), etc.}

\item{Newey}{Logical, whether the t-values are adjusted for heteroskedasticity and
auto-correlations (Newey and West, 1987).}

\item{GroupN}{Single numeric, how many portfolios are formed on each cross-section.}

\item{Detail}{Logical, whether show the High-Low (zero-investment / difference) portfolio.}
}
\value{
A data.table contains average value of \code{LHS} within each portfolio and the
t-statistics.
}
\description{
The objective of portfolio analysis is to examine the cross-sectional relation
between two or more variables.
}
\details{
At first we sort variable \code{RHS} and generate its rank to construct
portfolio, then we calculate the average value of \code{LHS} within each portfolio
for each period. Finally we examine whether these average values is statistically
distinguishable from zero.
}
\examples{
crs_port(DT = stock,
         LHS = "return",
         RHS = "size",
         Date = "month",
         Weight = "size",
         Risk = c("factor1", "factor2"),
         Newey = TRUE,
         GroupN = 10,
         Detail = FALSE)

}
\seealso{
Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
