% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_double.R
\name{crs_port2}
\alias{crs_port2}
\title{Double Sort Portfolio Analysis}
\usage{
crs_port2(
  DT,
  LHS,
  Control,
  RHS,
  Date,
  Weight = NULL,
  Risk = NULL,
  Newey = T,
  Independent = F,
  GroupN = c(3, 5)
)
}
\arguments{
\item{DT}{A data.frame or data.table, contains all data.}

\item{LHS}{Single character, the name of outcome / independent variable.}

\item{Control}{Single character, the name of first sort variable.}

\item{RHS}{Single character, the name of second sort variable.}

\item{Date}{Single character, variable name which indicate cross-section.}

\item{Weight}{Single character, variable name which used to calculate value-weighted mean.}

\item{Risk}{Character vector, risk adjust using factors like Fama-French (MKT, SMB, HML), etc.}

\item{Newey}{Logical, whether the t-values are adjusted for heteroskedasticity and
auto-correlations (Newey and West, 1987).}

\item{Independent}{Logical, whether use a unconditional sort.}

\item{GroupN}{Numeric vector of length 2, how many portfolios are formed on each cross-section.}
}
\value{
A data.table contains average value of \code{LHS} within each portfolio and the
t-statistics.
}
\description{
The objective of portfolio analysis is to examine the cross-sectional relation
between two or more variables.
}
\details{
At first we sort variable \code{Control} and generate its rank to construct
initial portfolio, then we sort \code{RHS} within (if \code{independent = F})
each initial portfolio and form final pofolio. Then we calculate the average value
of \code{LHS} within each portfolio for each period. Finally we examine whether
these average values is statistically distinguishable from zero.
}
\examples{
crs_port2(DT = stock,
          LHS = "return",
          Control = "explain",
          RHS = "size",
          Date = "month",
          Weight = "size",
          Risk = c("factor1", "factor2"),
          Newey = TRUE,
          Independent = FALSE,
          GroupN = c(3,5))

}
\seealso{
Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
