% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_regression.R
\name{crs_reg}
\alias{crs_reg}
\title{Cross-section Regression}
\usage{
crs_reg(DT, Y, Xs, Date)
}
\arguments{
\item{DT}{A data.table or data.frame contains variables.}

\item{Y}{Single character, the name of dependent / explained variable.}

\item{Xs}{Character vector, the name of independent / explanatory variable and control variables.}

\item{Date}{Single character, variable name which indicate cross-section.}
}
\value{
A data.table contains regression coefficients and t-value (in the bracket). Currently we
can't run a lot of regressions at the same time.
}
\description{
Run a Fama-MacBeth (1973) cross-section regression and summary the coefficients.
}
\details{
Every \strong{Date} (a cross-section) we run a OLS regression like \code{Y ~ X1 + X2}
(including intercept), the result of \code{crs_reg} summary average of the coefficients.
All the t-values are adjusted for heteroskedasticity and auto-correlations
(Newey and West, 1987).
}
\examples{
crs_reg(DT = stock, Y = "return", Xs = c("explain", "size"), Date = "month")

}
\seealso{
Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
