% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{crs_sum}
\alias{crs_sum}
\title{Cross-section Summary}
\usage{
crs_sum(DT, Date, NoUse, Use, Details = FALSE)
}
\arguments{
\item{DT}{A data.frame or data.table, contains all data.}

\item{Date}{A single character, variable name which indicate cross-section.}

\item{NoUse}{Character vector, some variables we don't care (including all character variables).}

\item{Use}{Character vector, reverse of NoUse, some variables we care. Can not use with
parameter NoUse.}

\item{Details}{Logical, should the function report the summary statistics every cross-section.}
}
\value{
A data.table have 15 columns, for example "Observations" / "Missings", etc.
Maybe we can use \code{t(...)} to make it more easy to present.
}
\description{
Cross-section description statistics is useful to understand the cross-sectional
distribution of the variables.
}
\details{
Robust summary because we average descriptive statistics, can effectively avoid the
influence of extreme values on descriptive statistics.
}
\examples{
crs_sum(DT = stock, Use = c("return", "size", "explain"), Date = "month")

crs_sum(DT = stock, NoUse = c("stkcd", "industry", "factor1", "factor2"), Date = "month")

}
\seealso{
Bali, T. ,  Engle, R. , &  Murray, S. . (2017). Empirical Asset Pricing: The Cross
Section of Stock Returns.
}
