% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{download_packages_electron}
\alias{download_packages_electron}
\title{Download packages for Electron app}
\usage{
download_packages_electron(
  project_path,
  app_bundle_path,
  r_version_minor,
  fallback_r_version,
  arch = "arm64"
)
}
\arguments{
\item{project_path}{Project path with renv.lock}

\item{app_bundle_path}{Path to the .app bundle}

\item{r_version_minor}{R version for package compatibility}

\item{fallback_r_version}{Fallback R version}

\item{arch}{Architecture}
}
\description{
Downloads R packages to the Electron app's Resources/library directory.
}
\keyword{internal}
