% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{download_packages_mac}
\alias{download_packages_mac}
\title{Download macOS binary packages}
\usage{
download_packages_mac(
  project_path,
  dist_dir,
  r_version_minor,
  fallback_r_version,
  arch = "arm64"
)
}
\arguments{
\item{project_path}{Path to project root}

\item{dist_dir}{Distribution directory}

\item{r_version_minor}{R major.minor version}

\item{fallback_r_version}{Fallback version}

\item{arch}{Architecture}
}
\value{
Invisible TRUE
}
\description{
Download macOS binary packages
}
\keyword{internal}
