% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_windows.R
\name{download_packages_windows}
\alias{download_packages_windows}
\title{Download Windows binary packages}
\usage{
download_packages_windows(
  project_path,
  dist_dir,
  r_version_minor,
  fallback_r_version
)
}
\arguments{
\item{project_path}{Path to project root}

\item{dist_dir}{Distribution directory}

\item{r_version_minor}{R major.minor version}

\item{fallback_r_version}{Fallback version}
}
\value{
Invisible TRUE
}
\description{
Download Windows binary packages
}
\keyword{internal}
