% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_na.R
\name{dropnas}
\alias{dropnas}
\title{Remove NA observations}
\usage{
dropnas(DT, Cols, KeepAllCols = T)
}
\arguments{
\item{DT}{A data.table or a data.frame.}

\item{Cols}{Character vector, contains the columns' name we want to remove NA values.}

\item{KeepAllCols}{Logical, if FALSE then the return value will only contain columns in \code{Cols}.
Equivalent to \code{select} and \code{na.omit} at the same time.}
}
\value{
A data.table without NA value (in \code{Cols}).
}
\description{
Remove NA values at the same time, in many columns.
}
\examples{
nrow(stock)
nrow(dropnas(stock, Cols = c("return", "size")))

dropnas(stock, Cols = c("return", "size"), KeepAllCols = FALSE)

}
