% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{extract_r_electron}
\alias{extract_r_electron}
\title{Extract R for Electron app}
\usage{
extract_r_electron(r_pkg_path, app_bundle_path)
}
\arguments{
\item{r_pkg_path}{Path to R .pkg file}

\item{app_bundle_path}{Path to the .app bundle}
}
\description{
Extracts R.framework into the Electron app's Frameworks directory.
}
\keyword{internal}
