% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{extract_r_mac}
\alias{extract_r_mac}
\title{Extract R for macOS from .pkg installer}
\usage{
extract_r_mac(r_pkg_path, dist_dir)
}
\arguments{
\item{r_pkg_path}{Path to R .pkg file}

\item{dist_dir}{Distribution directory}
}
\value{
Invisible TRUE if successful
}
\description{
Extract R for macOS from .pkg installer
}
\keyword{internal}
