% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{extract_r_mac_app}
\alias{extract_r_mac_app}
\title{Extract R for macOS app bundle from .pkg installer}
\usage{
extract_r_mac_app(r_pkg_path, dist_dir, app_name)
}
\arguments{
\item{r_pkg_path}{Path to R .pkg file}

\item{dist_dir}{Distribution directory}

\item{app_name}{Name of the app}
}
\value{
Invisible TRUE if successful
}
\description{
Extract R for macOS app bundle from .pkg installer
}
\keyword{internal}
