% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_windows.R
\name{extract_r_windows}
\alias{extract_r_windows}
\title{Extract R for Windows from installer}
\usage{
extract_r_windows(r_exe_path, dist_dir, deps)
}
\arguments{
\item{r_exe_path}{Path to R installer}

\item{dist_dir}{Distribution directory}

\item{deps}{System dependencies list}
}
\value{
Invisible TRUE if successful
}
\description{
Extract R for Windows from installer
}
\keyword{internal}
