% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{fix_r_framework_paths}
\alias{fix_r_framework_paths}
\title{Fix R.framework paths to make it relocatable}
\usage{
fix_r_framework_paths(framework_path)
}
\arguments{
\item{framework_path}{Path to R.framework}
}
\value{
Invisible TRUE
}
\description{
R.framework contains hardcoded paths to /Library/Frameworks/R.framework.
This function fixes scripts and symlinks to work from any location.
}
\keyword{internal}
