% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{fix_r_framework_paths_electron}
\alias{fix_r_framework_paths_electron}
\title{Fix R.framework paths for Electron app portability}
\usage{
fix_r_framework_paths_electron(r_framework_path)
}
\arguments{
\item{r_framework_path}{Path to R.framework}
}
\description{
Fix R.framework paths for Electron app portability
}
\keyword{internal}
