% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cran_binary_repo}
\alias{get_cran_binary_repo}
\title{Get CRAN binary repository URL}
\usage{
get_cran_binary_repo(r_version, target = c("windows", "mac"))
}
\arguments{
\item{r_version}{R major.minor version (e.g., "4.5")}

\item{target}{Target platform}
}
\value{
URL string
}
\description{
Get CRAN binary repository URL
}
\keyword{internal}
