% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{get_cran_binary_repo_mac}
\alias{get_cran_binary_repo_mac}
\title{Get CRAN binary repository URL for macOS}
\usage{
get_cran_binary_repo_mac(r_version, arch = "arm64")
}
\arguments{
\item{r_version}{R major.minor version}

\item{arch}{Architecture}
}
\value{
URL string
}
\description{
Get CRAN binary repository URL for macOS
}
\keyword{internal}
