% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_r_download_url}
\alias{get_r_download_url}
\title{Get R download URL}
\usage{
get_r_download_url(version, target = c("windows", "mac"))
}
\arguments{
\item{version}{R version string (e.g., "4.5.2")}

\item{target}{Target platform}
}
\value{
URL string
}
\description{
Get R download URL
}
\keyword{internal}
