% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{get_r_download_url_mac}
\alias{get_r_download_url_mac}
\title{Get R download URL for macOS}
\usage{
get_r_download_url_mac(version, arch = "arm64")
}
\arguments{
\item{version}{R version string}

\item{arch}{Architecture (arm64 or x86_64)}
}
\value{
URL string
}
\description{
Get R download URL for macOS
}
\keyword{internal}
