% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trick_date.R
\name{\%ym-\%}
\alias{\%ym-\%}
\title{Year-Month Shifter}
\usage{
YYYYMM \%ym-\% Shift
}
\arguments{
\item{YYYYMM}{Character vector in YYYY-MM (e.g. 2021-06, 1999-09) or YYYY-MM-DD format
(e.g. 2020-09-26, 1999-06-16)}

\item{Shift}{Integer, how many month ago (positive integer) or later (negative integer).}
}
\value{
A character vector in YYYY-MM or YYYY-MM-DD format (depends on input format,
and DD will not change).
}
\description{
In empirical research, to avoid Former-Bias, we need to adjust the date
frequently to stagger the time of different variables.
}
\examples{
# 5 months after January 2021
"2021-01" \%ym-\% -5

# 12 months before September 26, 2021
"2021-09-26" \%ym-\% 12

# The input can be a character vector
c("2021-01-15", "2021-02-25") \%ym-\% 5

}
