% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portR-package.R
\docType{package}
\name{portR-package}
\alias{portR-package}
\alias{portR}
\title{portR: Build Portable R Application Distributions}
\description{
portR provides tools to create portable distributions of R Shiny
applications that can run without requiring R installation on the
target machine. It supports building for Windows and macOS.
}
\section{Main Functions}{

\describe{
\item{\code{\link{build_windows}}}{Build a portable Windows distribution}
\item{\code{\link{build_mac}}}{Build a portable macOS distribution}
\item{\code{\link{check_build_requirements}}}{Check system requirements}
}
}

\section{Requirements}{

Projects must use renv for dependency management. The renv.lock file
is used to determine which packages to include in the distribution.
}

\author{
\strong{Maintainer}: Alessandro Gerada \email{alessandro.gerada@liverpool.ac.uk}

}
\keyword{internal}
