% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_renv_lock}
\alias{read_renv_lock}
\title{Read renv.lock file}
\usage{
read_renv_lock(project_path)
}
\arguments{
\item{project_path}{Path to the project root}
}
\value{
List containing lockfile contents
}
\description{
Read renv.lock file
}
\keyword{internal}
