% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{rebrand_electron_app}
\alias{rebrand_electron_app}
\title{Rebrand Electron app with custom name}
\usage{
rebrand_electron_app(app_bundle_path, app_name)
}
\arguments{
\item{app_bundle_path}{Path to the .app bundle}

\item{app_name}{New application name}
}
\description{
Updates Info.plist files and renames executables for rebranding.
}
\keyword{internal}
