% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_dist_directory}
\alias{setup_dist_directory}
\title{Create directory structure for distribution}
\usage{
setup_dist_directory(dist_dir, target = c("windows", "mac"))
}
\arguments{
\item{dist_dir}{Path to distribution directory}

\item{target}{Target platform}
}
\value{
Invisible TRUE
}
\description{
Create directory structure for distribution
}
\keyword{internal}
