% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mac.R
\name{setup_mac_app_bundle}
\alias{setup_mac_app_bundle}
\title{Setup macOS application bundle structure}
\usage{
setup_mac_app_bundle(dist_dir, app_name)
}
\arguments{
\item{dist_dir}{Distribution directory}

\item{app_name}{Name of the app}
}
\value{
Invisible TRUE
}
\description{
Setup macOS application bundle structure
}
\keyword{internal}
