% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stock.R
\docType{data}
\name{stock}
\alias{stock}
\title{Ordinary stock trading data (monthly)}
\format{
A data.table with 36000 rows and 9 columns:
\describe{
  \item{stkcd}{integer, [1, 100], stock indicator}
  \item{year}{character, [1992, 2021], year indicator}
  \item{month}{character, [1992-01, 2021-12], YYYY-MM format, month indicator}
  \item{industry}{character, [A, J], industry indicator}
  \item{return}{double, [-0.1, 0.1], monthly return, with 360 NAs}
  \item{explain}{double, [-5, 4], expalin variable, with 480 NAs}
  \item{size}{double, [0, 1], stock market value}
  \item{factor1}{double, [-3, 3], monthly risk-factor, with 240 NAs}
  \item{factor2}{double, [0, 1], monthly risk-factor, with 240 NAs}
}
}
\usage{
stock
}
\description{
A dataset containing the stock monthly returns and other attributes of 100
stocks. All variables of this data is generated randomly.
}
\keyword{datasets}
