% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trick_names.R
\name{use}
\alias{use}
\title{Colnames Shortcut}
\usage{
use(DT, NoUse, Regex = FALSE)
}
\arguments{
\item{DT}{Anything with names, usually a data.table.}

\item{NoUse}{Character vector, the names we don't use. Notice that when
\code{Regex = TRUE} the argument should be regular expression.}

\item{Regex}{Logical, if TRUE then we will use regular expression to filter
undesired variable names.}
}
\value{
A character vector without undesired variable names.
}
\description{
The package always use column names, when there are many variables in
the data, this function will reduce the workload.
}
\examples{
names(stock)
# we want to get all variable names without "factor1" and "factor2"
use(DT = stock, NoUse = c("factor1", "factor2"), Regex = FALSE)
# which is equivalent to following
use(DT = stock, NoUse = "^fac", Regex = TRUE)

}
