% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extreme.R
\name{winsor}
\alias{winsor}
\title{Winsorize vector}
\usage{
winsor(x, Probs, Cutpoints, Integer = F)
}
\arguments{
\item{x}{A numeric vector, always contains some extreme values.}

\item{Probs}{Numeric vector of length 2, the level of winsorize (must smaller than 1
and bigger than 0).}

\item{Cutpoints}{Numeric vector of length 2, the range of input x. Sets the values of a given variable
that are above or below a certain Cutpoint to that Cutpoint.}

\item{Integer}{Logical, if TRUE, the Integer variable will also be \strong{winsorized}; if
False, \code{winsor} just return the input itself.}
}
\value{
A numeric vector with less extreme values.
}
\description{
Make the value of a variable within a given level, often used to eliminate extreme values.
}
\examples{
summary(stock$return)
summary(winsor(stock$return, c(0.05, 0.95)))

identical(winsor(stock$industry), stock$industry)

}
